refactor config handling
