+++ config/config.example.js
