+++ package.json
@@ -25,6 +25,7 @@
     "format:scss": "stylelint --fix ./src",
     "lint": "eslint ./src",
     "lint:scss": "stylelint ./src",
+    "postinstall": "./node_modules/.bin/cpy \"./config/config.example.js\" \"./config\" --no-overwrite --rename=\"config.js\"",
     "precommit": "lint-staged",
     "prepush": "npm test",
     "start": "webpack-dev-server --hot --config webpack/dev.config.js",
@@ -82,6 +83,7 @@
     "babel-preset-env": "1.6.1",
     "babel-preset-react": "6.24.1",
     "babel-preset-stage-2": "6.24.1",
+    "cpy-cli": "1.0.1",
     "cross-env": "5.1.4",
     "css-loader": "0.28.11",
     "enzyme": "3.3.0",
