+++ src/utils/get-board-name.js
@@ -1,5 +1,5 @@
 import { findIndex, get } from 'lodash'
-import Config from '../../config'
+import Config from '../../config/config'
 
 export const getBoardName = (name = '') => {
   const index = findIndex(Config.boards, b => b.board === name)
