+++ src/utils/get-company-member.js
@@ -1,5 +1,5 @@
 import { find } from 'lodash'
-import { company_member as companyMember } from '../../config'
+import { company_member as companyMember } from '../../config/config'
 
 const getCompanyMember = members => find(members, { username: companyMember })
